Вынос логирующей мидлвары в отдельный файл
