+++ hw12_13_14_15_calendar/internal/server/httpserver/server.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -33,8 +32,11 @@ func newServer(app app.App, logger logger.Logger) Server {
 
 func (s *server) Start(addr string) error {
 	s.srv = &http.Server{
-		Addr:    addr,
-		Handler: s.router,
+		Addr:         addr,
+		Handler:      s.router,
+		WriteTimeout: time.Second * 15,
+		ReadTimeout:  time.Second * 15,
+		IdleTimeout:  time.Second * 60,
 	}
 	err := s.srv.ListenAndServe()
 	if errors.Is(err, http.ErrServerClosed) {
@@ -51,49 +53,9 @@ func (s *server) Stop(ctx context.Context) error {
 	return nil
 }
 
-func (s *server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	s.router.ServeHTTP(w, r)
-}
-
 func (s *server) configureRouter() {
-	s.router.Use(s.loggingMiddleware)
-	s.router.HandleFunc("/hello", s.handleHello).Methods("GET")
-}
-
-func (s *server) loggingMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		rw := &responseWriter{w, http.StatusOK}
-		next.ServeHTTP(rw, r)
-
-		s.logger.Info(
-			fmt.Sprintf("%s %s %s %s %d %s %s",
-				requestAddr(r),
-				r.Method,
-				r.RequestURI,
-				r.Proto,
-				rw.code,
-				latency(start),
-				userAgent(r),
-			))
-	})
-}
-
-func requestAddr(r *http.Request) string {
-	return strings.Split(r.RemoteAddr, ":")[0]
-}
-
-func userAgent(r *http.Request) string {
-	userAgents := r.Header["User-Agent"]
-	if len(userAgents) > 0 {
-		return "\"" + userAgents[0] + "\""
-	}
-	return ""
-}
-
-func latency(start time.Time) string {
-	return fmt.Sprintf("%dms", time.Since(start).Milliseconds())
+	s.router.Use(loggingMiddleware(s.logger))
+	s.router.HandleFunc("/hello", s.handleHello).Methods(http.MethodGet)
 }
 
 func (s *server) handleHello(w http.ResponseWriter, _ *http.Request) {
